+++ tilings/tiling.py
@@ -993,13 +993,20 @@ class Tiling(CombinatorialClass):
             or all(c in cells for c in chain.from_iterable(r.pos for r in req))
         )
         assumptions = tuple(
-            ass
-            for ass in self._assumptions
-            if (factors and ass.gps[0].pos[0] in cells)
-            or all(c in cells for c in chain.from_iterable(gp.pos for gp in ass.gps))
+            ass.__class__(
+                gp
+                for gp in ass.gps
+                if (factors and gp.pos[0] in cells) or all(c in cells for c in gp.pos)
+            )
+            for ass in self.assumptions
         ) + tuple(add_assumptions)
+        # TODO: check sum/skew assumptions
         return self.__class__(
-            obstructions, requirements, assumptions, simplify=False, sorted_input=True
+            obstructions,
+            requirements,
+            set(ass for ass in assumptions if ass.gps),
+            simplify=False,
+            sorted_input=True,
         )
 
     def find_factors(self, interleaving: str = "none") -> Tuple["Tiling", ...]:
@@ -1174,6 +1181,17 @@ class Tiling(CombinatorialClass):
             raise ValueError(f"following assumption not on tiling: '{assumption}'")
         return "k_{}".format(idx)
 
+    def get_assumption(self, parameter: str) -> TrackingAssumption:
+        idx = parameter.split("_")[1]
+        return self.assumptions[int(idx)]
+
+    def get_minimum_value(self, parameter: str) -> int:
+        """
+        Return the minimum value that can be taken by the parameter.
+        """
+        assumption = self.get_assumption(parameter)
+        return min(assumption.get_value(gp) for gp in self.minimal_gridded_perms())
+
     def maximum_length_of_minimum_gridded_perm(self) -> int:
         """Returns the maximum length of the minimum gridded permutation that
         can be gridded on the tiling.
@@ -1209,18 +1227,12 @@ class Tiling(CombinatorialClass):
         )
 
     def objects_of_size(self, n: int, **parameters: int) -> Iterator[GriddedPerm]:
-        if not parameters:
-            yield from self.gridded_perms_of_length(n)
-        else:
-            assert set(self.extra_parameters) == set(
-                parameters
-            ), f"{self.extra_parameters, set(parameters)}"
-            for gp in self.gridded_perms_of_length(n):
-                if all(
-                    ass.get_value(gp) == parameters[k]
-                    for k, ass in zip(self.extra_parameters, self._assumptions)
-                ):
-                    yield gp
+        for gp in self.gridded_perms_of_length(n):
+            if all(
+                self.get_assumption(k).get_value(gp) == val
+                for k, val in parameters.items()
+            ):
+                yield gp
 
     def gridded_perms_of_length(self, length: int) -> Iterator[GriddedPerm]:
         for gp in self.gridded_perms(maxlen=length):